backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
PyMySQL==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
bcrypt==3.2.0
stripe==2.60.0
Flask-Mail==0.9.1
email-validator==1.1.3 